Mudança para aparecer os valores correto dos pokemons

+++ app/src/main/java/com/example/pokedexhacksprint/common/data/Entity.kt
@@ -2,14 +2,21 @@ package com.example.pokedexhacksprint.common.data
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import com.example.pokedexapp.PokemonDto
 
 @Entity(tableName = "pokemon_table")
 data class PokemonEntity(
     @PrimaryKey val id: Int,
     val name: String,
-    val height: Int,
-    val weight: Int,
+    val height: Double,
+    val weight: Double,
     val types: String,
     val stats: String,
     val frontDefault: String
+
+
+
 )
+
+
+
